requirements.txt
streamlit>=1.31.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
opencv-python>=4.8.0
streamlit-image-coordinates>=0.1.6